+++ Decsys/ClientApp/src/app/screens/admin/EditorScreen/PureEditorScreen.js
@@ -3,12 +3,12 @@ import PropTypes from "prop-types";
 import EditorBar from "components/EditorBar";
 import { Grid, Cell } from "styled-css-grid";
 import EditorPageList from "components/EditorPageList";
-import { EmptyState } from "components/ui";
+import { EmptyState } from "components/core";
 import { FileAlt } from "styled-icons/fa-solid";
 import { Box, colorVariant } from "@smooth-ui/core-sc";
 import ComponentEditor from "components/ComponentEditor";
 import ImageUpload from "components/ComponentEditor/ImageUpload";
-import { getComponent } from "utils/component-utils";
+import { getComponent } from "services/page-items";
 import SurveyPageBody from "components/SurveyPage/Body";
 
 const PureEditorScreen = ({
