+++ Decsys/ClientApp/src/app/screens/admin/SurveysScreen/AddSurveyModal.js
@@ -14,8 +14,7 @@ import {
   FileImport,
   ExclamationCircle
 } from "styled-icons/fa-solid";
-import { FlexBox } from "components/ui";
-import ConfirmModal from "components/ui/ConfirmModal";
+import { FlexBox, ConfirmModal } from "components/core";
 
 const AddSurveyModal = ({
   modalState,
