+++ Decsys/ClientApp/src/app/screens/admin/SurveysScreen/PureSurveysScreen.js
@@ -2,9 +2,8 @@ import React from "react";
 import PropTypes from "prop-types";
 import { Typography, Button, Alert, Box } from "@smooth-ui/core-sc";
 import { List, PlusCircle, InfoCircle } from "styled-icons/fa-solid";
-import { Container, FlexBox, EmptyState } from "components/ui";
+import { Container, FlexBox, EmptyState, useModal } from "components/core";
 import SurveyList from "components/SurveyList";
-import { useModal } from "components/ui/ConfirmModal";
 import AboutLink from "components/AboutLink";
 import AppBar from "components/AppBar";
 import AddSurveyModal from "./AddSurveyModal";
